parquet: avoid potential long truncation
Was using a long where I should have had an int.
This change prevents a [lossy-conversions] warning from Java 21.

+++ parquet/src/main/uk/ac/starlink/parquet/SequentialParquetStarTable.java
@@ -126,7 +126,7 @@ public class SequentialParquetStarTable extends ParquetStarTable {
         }
 
         public long splittableSize() {
-            int nr = 0;
+            long nr = 0;
             for ( int ib = iblock_ + 1; ib < iblockEnd_; ib++ ) {
                 nr += blockSizes_[ ib ];
             }
